# docs/swagger.yaml
basePath: /api
definitions:
  models.Comment:
    properties:
      author_id:
        type: integer
      author_name:
        type: string
      can_delete:
        type: boolean
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      post_id:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: API для форума
  title: Forum Service API
  version: "1.0"
paths:
  /chat:
    get:
      description: Возвращает список всех сообщений в общем чате.
      produces:
      - application/json
      responses:
        "200":
          description: Returns list of messages
          schema:
            items:
              type: object
            type: array
        "500":
          description: ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить все сообщения чата
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Создаёт новое сообщение в общем чате. Доступно только авторизованным
        пользователям.
      parameters:
      - description: Данные для создания сообщения
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Returns created message
          schema:
            type: object
        "400":
          description: неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: пользователь не аутентифицирован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать сообщение в чате
      tags:
      - chat
  /comments:
    post:
      consumes:
      - application/json
      description: Создаёт новый комментарий к посту. Доступно только авторизованным
        пользователям.
      parameters:
      - description: Данные для создания комментария
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: пользователь не аутентифицирован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать новый комментарий
      tags:
      - comments
  /comments/{id}:
    delete:
      description: Удаляет комментарий. Доступно только автору комментария или администратору.
      parameters:
      - description: ID комментария
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Неверный ID комментария
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: пользователь не аутентифицирован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: нет прав для удаления этого комментария
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить комментарий
      tags:
      - comments
  /comments/chat:
    post:
      consumes:
      - application/json
      description: Создаёт новое сообщение в общем чате. Доступно только авторизованным
        пользователям.
      parameters:
      - description: Данные для создания сообщения
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка при создании сообщения
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать сообщение в чате
      tags:
      - comments
...
schemes:
- http
- https
swagger: "2.0"